server/requirements.txt
# Flask
Flask==0.12.2
flask-cors==3.0.3
PyJWT==1.5.3

# Database
Flask-SQLAlchemy==2.1
Flask-Migrate==2.1.1
psycopg2==2.7.3.1

# Amazon
bottlenose==1.1.3
beautifulsoup4==4.4.0
lxml==4.0.0

# Testing
pytest-flask==0.10.0

# Deployment
gunicorn==19.7